requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
sentence-transformers
transformers
faiss-cpu
Pillow
jinja2
aiofiles
flask
flask-wtf
flask-limiter




